Adds the `listener` class
The `listener` class is a threaded class that receives health messages
and renders the updates of these messages to the terminal.

It has the method `std::string construc_string(messages::HealthMessage)`
that constructs a string in the format of *<name> - <status> {<message>}*.

It also has the `void render()` function that clears the terminal first and
then prints all the status messages.

The class also has the `bool update` method that indicates whether a new
render has to be executed.